requirements.txt
whatsapp-python-sdk
aiohttp
aiofiles
pillow
pytube